[update] add args to save vedio in some func
