+++ src/main.rs
@@ -1,5 +1,7 @@
-mod stream;
+mod resources;
 mod folder;
+use resources::{FrameTools, FrameDetection};
+
 
 
 fn main() ->Result<(),opencv::Error>{
@@ -8,14 +10,22 @@ fn main() ->Result<(),opencv::Error>{
     for folder_name in folder_names{
         folder::set_folder(folder_name);
     }
-    
-    // Use Camera
-    let mut video_stream = stream::Stream::from_video("328_1687706104.mp4");
-    video_stream.moving_object_detection(3, 20, 30)?;
-    let mut my_camera = stream::Stream::from_camera();
-    // my_camera.moving_object_detection(6, 60, 30)?;
-    my_camera.face_detection()?;
-    my_camera.body_detection()?;
+
+    // Use Pics as Frame
+    let mut baoguo_pic = resources::Frame::get_from_img("TeacherBaoguo.jpeg")?;
+    baoguo_pic.face_detection()?;
+    baoguo_pic.show("Baoguo")?;
+    baoguo_pic.save_as_img("for_readme").unwrap();
+
+    // Use Vedio as Stream
+    let mut video_stream = resources::Stream::from_video("328_1687706104.mp4");
+    video_stream.moving_object_detection(3, 20, 30, false, "moving_vedio.mp4")?;
+
+    // // Use Camera as Stream
+    // let mut my_camera = resources::Stream::from_camera();
+    // my_camera.moving_object_detection(6, 60, 60,true, "moving_cam.mp4")?;
+    // my_camera.face_detection(true, "face_cam.mp4")?;
+    // my_camera.body_detection(true,"body_cam.mp4")?;
     // my_camera.camera()?;
     // my_camera.capture_frame()?;
