+++ src/resources/consts.rs
@@ -1,0 +1,2 @@
+pub const VEDIO: i32 = 0;
+pub const CAMERA: i32 = 1;
